# spec/openapi.yaml
openapi: 3.1.0
info:
  title: ATOM Sovereign Agent Runtime API
  description: |
    ATOM.0-alpha — Sovereign Recursive Agent Architecture.
    
    This API exposes the runtime surface for mission control, effect dispatch,
    capability management, and evidence retrieval. All mutation paths traverse
    the sovereign kernel boundary (authorize → commit).
    
    ## Authentication

    Every route except `GET /health` and `GET /ready` requires
    `Authorization: Bearer <token>`. The token is a provisioned operator
    secret (opaque, minimum 16 bytes), not a JWT: it proves the caller may
    talk to the daemon, never *who* they are. Missing or wrong credentials
    yield `401` with a `WWW-Authenticate: Bearer realm="atom"` challenge.
    The daemon refuses to start without a token unless `--no-auth` is passed
    explicitly (loopback development only). Auth is transport only — it never
    substitutes for capability grants or approvals.
    
    ## Idempotency
    All mutation endpoints accept `Idempotency-Key` header.
    Duplicate requests with the same key return the original response.
    
    ## Error Format
    Errors follow RFC 9457 (Problem Details):
    ```json
    {
      "type": "https://atom.dev/errors/kernel-denied",
      "title": "Kernel Authorization Denied",
      "status": 403,
      "detail": "Capability grant does not cover requested operation",
      "instance": "/missions/abc123/effects/def456"
    }
    ```
  version: 4.0.0-alpha
  license:
...
servers:
...
paths:
...
  /approvals:
    get:
      tags: [approvals]
      summary: List approval grants
      operationId: listApprovals
      description: |
        Lists owner approval grants and whether each has been redeemed. An
        exact-effect approval authorises exactly one crossing (AUT-003).
      responses:
        '200':
          description: Approval grant list
          content:
            application/json:
              schema:
                type: object
                properties:
                  approvals:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApprovalGrant'
    post:
      tags: [approvals]
      summary: Issue an approval grant
      operationId: issueApproval
      description: |
        Records an owner decision authorising either one exact effect digest or
        a bounded capability scope. This is the human gate in front of
        `POST /host/commit`: obtain the digest from `POST /host/plan`, approve
        it here, then commit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalGrant'
      responses:
        '201':
          description: Approval recorded durably
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalGrant'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /host/commit:
    post:
      tags: [host]
      summary: Commit a declared host mutation
      operationId: commitHostOp
      description: |
        Redeems a durable approval grant for the plan's exact effect digest,
        re-validates authority and the resource witness at the boundary, issues
        a one-shot CommitPermit bound to this connector identity, and only then
        asks the privilege broker to admit the operation (KRN-002, EFX-004,
        AUT-003).

        The one-shot nonce is burned durably before the response returns, so the
        guarantee survives a restart. The approval is spent in the same step: an
        exact-effect approval authorises exactly one crossing.

        Refusals: no approval for this digest (400), resource witness drifted
        since planning (409), plan already committed (409), privilege boundary
        denied the crossing (409), host surface disabled (503).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostCommitRequest'
      responses:
        '200':
          description: The host was changed through the privilege boundary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostCommitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Witness drift, spent plan, or boundary denial
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Host mutation surface is disabled (no --host-root)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Operator bearer token (`Authorization: Bearer <token>`). Required on
        every operation except `GET /health` and `GET /ready`. Transport only:
        it never substitutes for capability grants or approvals.
...
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    HostCommitRequest:
      type: object
      additionalProperties: false
      required: [plan_id]
      properties:
        plan_id:
          type: string
    HostCommitResponse:
      type: object
      required:
        - plan_id
        - permit_id
        - one_shot_nonce
        - approval_id
        - effect_digest
        - outcome
        - observation_id
        - state
      properties:
        plan_id:
          type: string
        permit_id:
          type: string
        one_shot_nonce:
          type: string
          description: Burned durably before this response returned.
        approval_id:
          type: string
        effect_digest:
          type: string
        outcome:
          type: string
          description: The executor's own account of what it did.
        observation_id:
          type: string
          description: The sealed evidence entry, tainted HOST_MUTATION_COMMITTED.
        state:
          const: COMMITTED
...
    ValidityInterval:
      type: object
      description: A half-open interval `[not_before, expires_at)`.
      required: [not_before, expires_at]
      properties:
        not_before:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    ApprovalScope:
      description: |
        The mutually exclusive scopes AUT-003 permits. An `effect` scope names one
        exact effect digest; a `capability` scope bounds an operation class.
        Effect and capability targets never cross-match.
      oneOf:
        - type: object
          additionalProperties: false
          required: [kind, effect_digest]
          properties:
            kind:
              const: effect
            effect_digest:
              type: string
        - type: object
          required: [kind]
          properties:
            kind:
              const: capability
      discriminator:
        propertyName: kind
    ApprovalGrant:
      type: object
      description: A durable owner decision that can authorise a crossing.
      required: [grant_id, approver_id, scope, validity]
      properties:
        grant_id:
          type: string
        approver_id:
          type: string
          description: The human or authority that approved the scope.
        scope:
          $ref: '#/components/schemas/ApprovalScope'
        validity:
          $ref: '#/components/schemas/ValidityInterval'
        revocation_state:
          type: string
          enum: [ACTIVE, REVOKED, EXPIRED]
        generation:
          type: integer
        redeemed:
          type: boolean
          description: Whether this approval has already been spent.
        attestation:
          type: object
          description: |
            Daemon attestation stapled at issue (P0-B). Proves the approval
            passed through this daemon's authenticated API unmodified since.
            Absent on pre-attestation records. Does not identify a human.
          properties:
            key_id:
              type: string
            signature:
              type: string
          required: [key_id, signature]
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'